The-Fature/requirement.txt
pymysql == 0.9.3
django-cors-headers == 3.1.1
itsdangerous == 1.1.0
djangorestframework == 3.10.3
